add new travis files

+++ tests/integration_tests/travis/jumspcale_install.sh
@@ -1,10 +1,13 @@
+if [ ${1} = "True" ]; then
+    echo "Jumpscale already installed in this machine."
+    exit
+fi
 # prepare the machine for jumpscale installation 
 apt-get install -y python3.5 python3.5-dev
 sudo rm -f /usr/bin/python
 sudo rm -f /usr/bin/python3
 sudo ln -s /usr/bin/python3.5 /usr/bin/python
 sudo ln -s /usr/bin/python3.5 /usr/bin/python3
-
 
 ## install jumpscale 
 ./.travis_ci/prepare.sh
